requirements.txt
streamlit
faiss-cpu
sentence-transformers
scikit-learn
bs4
requests
google-generativeai 
